requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyPDF2
langchain
tiktoken
faiss-cpu
langchain-openai
